# Add timing cache to accelerate consequent `.engine` export