<!--
Thank you 🙏 for your contribution to [Ultralytics](https://www.ultralytics.com/) 🚀! Your effort in enhancing our repositories is greatly appreciated. To streamline the process and assist us in integrating your Pull Request (PR) effectively, please follow these steps:

1. Check for Existing Contributions: Before submitting, kindly explore existing PRs to ensure your contribution is unique and complementary.
2. Link Related Issues: If your PR addresses an open issue, please link it in your submission. This helps us better understand the context and impact of your contribution.
3. Elaborate Your Changes: Clearly articulate the purpose of your PR. Whether it's a bug fix or a new feature, a detailed description aids in a smoother integration process.
4. Ultralytics Contributor License Agreement (CLA): To uphold the quality and integrity of our project, we require all contributors to sign the CLA. Please confirm your agreement by commenting below:

    I have read the CLA Document and I sign the CLA

For more detailed guidance and best practices on contributing, refer to our ✅ [Contributing Guide](https://docs.ultralytics.com/help/contributing/). Your adherence to these guidelines ensures a faster and more effective review process.
--->
It's time consuming to export `.engine` format model with `--half` option. TensorRT provides an option to use timing cache to accelerate the consequent export process. This PR add this option to the `export.py`.

Export `yolov5m.pt` without timing cache take 565.4s:

```bash
python export.py --weights weights/yolov5m.pt --include engine --opset 17 --half --device 0
```

```bash
[10/25/2024-14:58:27] [TRT] [I] [MemUsageStats] Peak memory usage of TRT CPU/GPU memory allocators: CPU 5 MiB, GPU 213 MiB
[10/25/2024-14:58:27] [TRT] [I] [MemUsageStats] Peak memory usage during Engine building and serialization: CPU: 3283 MiB
TensorRT: export success ✅ 565.4s, saved as weights/yolov5m.engine (42.3 MB)

Export complete (566.4s)
```

And shrink to 11.5s with timing cache (the second run):

```bash
python export.py --weights weights/yolov5m.pt --include engine --opset 17 --half --device 0 --cache runs/timing.cache
```

```bash
[10/25/2024-15:08:19] [TRT] [I] [MemUsageStats] Peak memory usage of TRT CPU/GPU memory allocators: CPU 5 MiB, GPU 44 MiB
[10/25/2024-15:08:19] [TRT] [I] [MemUsageStats] Peak memory usage during Engine building and serialization: CPU: 4374 MiB
[10/25/2024-15:08:19] [TRT] [I] Serialized 27 bytes of code generator cache.
[10/25/2024-15:08:19] [TRT] [I] Serialized 14672 timing cache entries
TensorRT: export success ✅ 11.5s, saved as weights/yolov5m.engine (42.3 MB)

Export complete (12.3s)
```

Reference for [`timing cache`](https://docs.nvidia.com/deeplearning/tensorrt/developer-guide/index.html#timing-cache)

## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
Enhancement of YOLOv5 export functionality by introducing TensorRT timing cache capability and adjusting image input handling.

### 📊 Key Changes
- **TensorRT Timing Cache**: Added an optional `cache` parameter to export the engine with a path for saving/loading the TensorRT timing cache.
- **Image Input Handling**: Modified the method to determine input image channels dynamically when setting up images.
- **Code Documentation**: Improved clarity and accuracy of comments regarding TensorRT and augmentation settings.

### 🎯 Purpose & Impact
- **Improved Export Performance**: By using the timing cache, TensorRT exports can become faster on subsequent runs, enhancing efficiency for users who frequently export models.
- **Enhanced Flexibility**: Dynamic input channels handling allows for more adaptive configurations, accommodating various model setups without manual tweaking.
- **Clarity and Maintenance**: Updated comments and documentation contribute to a more understandable and maintainable codebase, aiding both current and future developers.